interfaces/requirements.txt
python-telegram-bot==21.4
streamlit==1.37.0
